fix: schema error and test updation flow (#5124)
