+++ packages/hoppscotch-common/src/platform/std/kernel-interceptors/extension/store.ts
@@ -89,15 +89,12 @@ export class KernelInterceptorExtensionStore extends Service {
     }
 
     const watcher = await Store.watch(STORE_NAMESPACE, SETTINGS_KEY)
-    watcher.on(
-      "change",
-      async ({ value }) => {
-        if (value) {
-          const storedData = value as StoredData
-          this.settings = storedData.settings
-        }
+    watcher.on("change", async ({ value }) => {
+      if (value) {
+        const storedData = value as StoredData
+        this.settings = storedData.settings
       }
-    )
+    })
   }
 
   public getExtensionVersion(): O.Option<ExtensionVersion> {
