+++ packages/hoppscotch-common/src/platform/std/kernel-interceptors/native/store.ts
@@ -74,15 +74,12 @@ export class KernelInterceptorNativeStore extends Service {
 
   private async setupWatchers() {
     const watcher = await Store.watch(STORE_NAMESPACE, STORE_KEYS.SETTINGS)
-    watcher.on(
-      "change",
-      async ({ value }) => {
-        if (value) {
-          const store = value as StoredData
-          this.domainSettings = new Map(Object.entries(store.domains))
-        }
+    watcher.on("change", async ({ value }) => {
+      if (value) {
+        const store = value as StoredData
+        this.domainSettings = new Map(Object.entries(store.domains))
       }
-    )
+    })
   }
 
   private async persistStore(): Promise<void> {
