+++ packages/hoppscotch-data/src/index.ts
@@ -9,3 +9,4 @@ export * from "./utils/collection"
 export * from "./utils/hawk"
 export * from "./utils/akamai-eg"
 export * from "./utils/jwt"
+export * from "./rest-request-response"
