+++ packages/hoppscotch-data/src/rest-request-response/original-request/index.ts
@@ -1,0 +1,39 @@
+import { InferredEntity, createVersionedEntity } from "verzod"
+import { z } from "zod"
+import V1_VERSION from "./v/1"
+import V2_VERSION from "./v/2"
+import V3_VERSION from "./v/3"
+import V4_VERSION from "./v/4"
+import V5_VERSION from "./v/5"
+
+const versionedObject = z.object({
+  // v is a stringified number
+  v: z.string().regex(/^\d+$/).transform(Number),
+})
+
+export const HoppRESTResponseOriginalRequest = createVersionedEntity({
+  latestVersion: 5,
+  versionMap: {
+    1: V1_VERSION,
+    2: V2_VERSION,
+    3: V3_VERSION,
+    4: V4_VERSION,
+    5: V5_VERSION,
+  },
+  getVersion(data) {
+    const versionCheck = versionedObject.safeParse(data)
+
+    if (versionCheck.success) return versionCheck.data.v
+
+    // For V1 we have to check the schema
+    const result = V1_VERSION.schema.safeParse(data)
+
+    return result.success ? 1 : null
+  },
+})
+
+export const HoppRESTResOriginalReqSchemaVersion = "5"
+
+export type HoppRESTResponseOriginalRequest = InferredEntity<
+  typeof HoppRESTResponseOriginalRequest
+>
