+++ packages/hoppscotch-selfhost-desktop/src/platform/environments/environments.platform.ts
@@ -96,11 +96,10 @@ async function loadUserEnvironments() {
 
         replaceEnvironments(
           formatedEnvironments.map((environment) => {
-            const parsedEnv =
-              entityReference(Environment).safeParse(environment)
+            const parsedEnv = Environment.safeParse(environment)
 
-            return parsedEnv.success
-              ? parsedEnv.data
+            return parsedEnv.type === "ok"
+              ? parsedEnv.value
               : {
                   ...environment,
                   v: EnvironmentSchemaVersion,
@@ -191,12 +190,24 @@ function setupUserEnvironmentUpdatedSubscription() {
         )
 
         if ((localIndex || localIndex == 0) && name) {
+          const environment = {
+            id,
+            name,
+            variables: JSON.parse(variables),
+          }
+
+          const parsedEnvResult = Environment.safeParse(environment)
+
+          const parsedEnv: Environment =
+            parsedEnvResult.type === "ok"
+              ? parsedEnvResult.value
+              : {
+                  ...environment,
+                  v: EnvironmentSchemaVersion,
+                }
+
           runDispatchWithOutSyncing(() => {
-            updateEnvironment(localIndex, {
-              id,
-              name,
-              variables: JSON.parse(variables),
-            })
+            updateEnvironment(localIndex, parsedEnv)
           })
         }
       }
